Extract shared nudge footprint geometry module
Move the keep-footprint tables, side midpoint naming, split-side endpoint
preservation, and midpoint Z override helpers into nudge_geometry. Both nudge
modes cut affected corners off the same footprint and insert midpoints on the
same sides, so that arithmetic belongs in one place.

Drop the leading underscore from the names that now cross a module boundary,
and rename the create_cells locals that would otherwise shadow
full_cell_footprint.

Emitted meshes are unchanged: five smoke configs stay byte identical.
